requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==1.16.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.11.3
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
